requirements.txt
PyMuPDF
langchain
openai
python-dotenv
langchain_community
pdfplumber
PyPDF4
chromadb
langchain_openai